path checking before param file write

+++ src/Core/Deployments/TerraformVarsFile.cs
@@ -27,16 +27,31 @@ namespace Modm.Deployments
 
         public async Task Write(IDictionary<string, object> parameters)
         {
-            var json = JsonSerializer.Serialize(parameters, new JsonSerializerOptions
+            if (Directory.Exists(destinationDirectory))
             {
-                Converters = { new DictionaryStringObjectJsonConverter() }
-            });
-            await File.WriteAllTextAsync(FullPath, json);
+                var json = JsonSerializer.Serialize(parameters, new JsonSerializerOptions
+                {
+                    Converters = { new DictionaryStringObjectJsonConverter() }
+                });
+
+                // Validate write permission and file creation
+                using (FileStream fs = new FileStream(FullPath, FileMode.Create, FileAccess.Write))
+                {
+                    using (StreamWriter writer = new StreamWriter(fs))
+                    {
+                        await writer.WriteAsync(json);
+                    }
+                }
+            }
+
         }
 
-        public Task Delete()
+        public async Task Delete()
         {
-            return Task.Run(() => File.Delete(FullPath));
+            if (File.Exists(FullPath))
+            {
+                await Task.Run(() => File.Delete(FullPath));
+            }
         }
     }
 }
